requirements.txt
pandas==2.0.3
requests==2.31.0
dbt-postgres==1.5.3
psycopg2-binary==2.9.7
pymongo==4.4.0
redis==5.0.0
pytest==7.4.2
